# mimirpb: Optimize RW2 Marshalling