#### What this PR does

This PR optimizes marshalling of RW2. Benchmark shows a `-17.06%` change to CPU, `-43.14%` change to memory, and removal of nearly all allocations (previously scaled based on series count; now fixed at 2, 99.99% reduction in the benchmark).

<details>
<summary>Full benchstat results (only "Marshal/RW2" benchmarks affected)</summary>

```
pkg: github.com/grafana/mimir/pkg/mimirpb
cpu: AMD Ryzen 7 PRO 5850U with Radeon Graphics
                                               │ before.txt  │          after-reverse.txt          │
                                               │   sec/op    │   sec/op     vs base                │
UnMarshal/Marshal/RW1-16                         34.21m ± 1%   35.27m ± 1%   +3.09% (p=0.002 n=10)
UnMarshal/Marshal/RW2-16                         20.99m ± 1%   17.41m ± 1%  -17.06% (p=0.000 n=10)
UnMarshal/Unmarshal/RW1/skipExemplars=true-16    20.63m ± 3%   21.89m ± 3%   +6.11% (p=0.000 n=10)
UnMarshal/Unmarshal/RW2/skipExemplars=true-16    24.22m ± 1%   25.52m ± 1%   +5.33% (p=0.000 n=10)
UnMarshal/Unmarshal/RW1/skipExemplars=false-16   23.63m ± 2%   24.77m ± 1%   +4.84% (p=0.000 n=10)
UnMarshal/Unmarshal/RW2/skipExemplars=false-16   25.98m ± 1%   27.90m ± 3%   +7.42% (p=0.000 n=10)
geomean                                          24.57m        24.88m        +1.23%

                                               │  before.txt  │          after-reverse.txt           │
                                               │     B/op     │     B/op      vs base                │
UnMarshal/Marshal/RW1-16                         25.58Mi ± 0%   25.58Mi ± 0%   -0.00% (p=0.025 n=10)
UnMarshal/Marshal/RW2-16                         30.78Mi ± 0%   17.50Mi ± 0%  -43.14% (p=0.000 n=10)
UnMarshal/Unmarshal/RW1/skipExemplars=true-16    54.48Mi ± 0%   54.48Mi ± 0%        ~ (p=0.739 n=10)
UnMarshal/Unmarshal/RW2/skipExemplars=true-16    31.29Mi ± 0%   31.29Mi ± 0%        ~ (p=0.218 n=10)
UnMarshal/Unmarshal/RW1/skipExemplars=false-16   59.05Mi ± 0%   59.05Mi ± 0%        ~ (p=0.289 n=10)
UnMarshal/Unmarshal/RW2/skipExemplars=false-16   32.82Mi ± 0%   32.82Mi ± 0%        ~ (p=0.739 n=10)
geomean                                          37.08Mi        33.75Mi        -8.98%

                                               │   before.txt   │           after-reverse.txt           │
                                               │   allocs/op    │  allocs/op   vs base                  │
UnMarshal/Marshal/RW1-16                             2.000 ± 0%    2.000 ± 0%        ~ (p=1.000 n=10) ¹
UnMarshal/Marshal/RW2-16                         20002.000 ± 0%    2.000 ± 0%  -99.99% (p=0.000 n=10)
UnMarshal/Unmarshal/RW1/skipExemplars=true-16       60.43k ± 0%   60.43k ± 0%        ~ (p=1.000 n=10) ¹
UnMarshal/Unmarshal/RW2/skipExemplars=true-16       50.03k ± 0%   50.03k ± 0%        ~ (p=1.000 n=10) ¹
UnMarshal/Unmarshal/RW1/skipExemplars=false-16      100.4k ± 0%   100.4k ± 0%        ~ (p=1.000 n=10) ¹
UnMarshal/Unmarshal/RW2/skipExemplars=false-16      60.03k ± 0%   60.03k ± 0%        ~ (p=1.000 n=10)
geomean                                             9.487k        2.044k       -78.46%
```

</details>

Namely we focus on marshalling labels references, removing the need for a memory buffer in the generated code.
It's not an original optimization. Upstream Prometheus does the same, as added in [this commit](https://github.com/prometheus/prometheus/pull/14395/commits/5dbe3c191961b228c3eae08a5d827271d590c745) by @npazosmendez.

[I've also included an alternative implementation in this commit](https://github.com/grafana/mimir/commit/4934c540710788066ae9eddb76b4ff078159fa38), that builds on the Prometheus version. It just marshalls the references completely inplace without the need for reversing anything. It's about 5% faster in the benchmark (no change to memory or allocations). But, the tradeoff is that it's completely custom logic written from scratch - the `reverse` algorithm retains some of the generated code, and is easier to diff/understand without needing a deep understanding of protobuf encoding.

#### Which issue(s) this PR fixes or relates to

Contrib https://github.com/grafana/mimir-squad/issues/2253

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry is not needed, please add the `changelog-not-needed` label to the PR.
- [ ] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.